Thought I would manage to get to the end of the constant functions region but there were a lot more CMDLETs than I anticipated.
That's usually not the problem however all of the CMDlets following Transform-Parameters are incredibly complex and need to be approached from a different angle / require a bit more of a review first.

It is also a friday and I don't have it in me to keep going while sun is out.